Add B+ tree implementation in Go
A new file has been added to the Algorithm/b_plus_tree directory which includes the implementations of a B+ tree in Go. This includes insertion of keys and splitting of leaf nodes when necessary. Further improvements may include defining how to print the tree structure and handling the insertion of the new leaf's first key into the parent.